+++ profiles/stac-server/outputs.tf
@@ -25,3 +25,7 @@ output "stac_server_ingest_sns_topic_arn" {
 output "stac_server_post_ingest_sns_topic_arn" {
   value = module.stac-server.stac_server_post_ingest_sns_topic_arn
 }
+
+output "stac_server_lambda_iam_role_arn" {
+  value = module.stac-server.stac_server_lambda_iam_role_arn
+}